Authorizations

+++ src/controllers/things/destroy.controller.js
@@ -1,8 +1,10 @@
+import { resources, statusCodes } from '../../configs';
 import { things as thingsService } from '../../services';
+import authorizations from './authorizations.config';
 
-/** 
+/**
  * Destroy thing record with primary key id
- * 
+ *
  * @memberof module:controllers/things
  * @param {Object} request - HTTP request object
  * @param {Object} response - HTTP response callback object
@@ -10,21 +12,36 @@ import { things as thingsService } from '../../services';
  */
 const destroyOne = async (request, response, next) => {
   try {
+    // Resource being created. We use this to check authorization later.
+    const THINGS = resources.THINGS;
+    // User instance requesting controller
+    const requestUser = request.user;
+
+    // Check if email matches req.user.email and set permission
+    const isAny = authorizations.can(requestUser.role).deleteAny(THINGS);
+
+    // Check if permission is grated
+    if (!isAny.granted) {
+      const error = new Error(`AUTHORIZATION ERROR: You are not authorized to ${request.method} resources on ${THINGS}.`);
+      error.statusCode = statusCodes.UNAUTHORIZED;
+      throw error;
+    }
+
     // Check we have request params to parse
     if (!request.params) {
       const error = new Error('CONTROLLER ERROR: Your destroy thing request did not contain any params.');
-      error.statusCode = 400;
+      error.statusCode = statusCodes.BAD_REQUEST;
       throw error;
     }
 
     // Parse request params
-    const { id } = request.params;
+    const id = Number(request.params.id);
 
     // Check we have an id to update
     if (!id) {
       console.log(id);
       const error = new Error('CONTROLLER ERROR: Your destroy thing request did not contain a thing id.');
-      error.statusCode = 400;
+      error.statusCode = statusCodes.BAD_REQUEST;
       throw error;
     }
 
@@ -34,13 +51,13 @@ const destroyOne = async (request, response, next) => {
     // Check we have row count to respond with
     if (!countRowsDestroyed) {
       const error = new Error(`CONTROLLER ERROR: Unable to destroy thing ${id} record.`);
-      error.statusCode = 500;
+      error.statusCode = statusCodes.INTERNAL_SERVER_ERROR;
       throw error;
     }
 
     // Destroy response
-    const responseObject = response.status(200).json({
-      status: 200,
+    const responseObject = response.status(statusCodes.OK).json({
+      status: statusCodes.OK,
       message: `SUCCESS: Destroyed thing ${id} record.`,
       count: countRowsDestroyed,
     });
@@ -51,9 +68,9 @@ const destroyOne = async (request, response, next) => {
   }
 };
 
-/** 
+/**
  * Destroy all things in the database
- * 
+ *
  * @memberof module:controllers/things
  * @param {Object} request - HTTP request object
  * @param {Object} response - HTTP response object callback
@@ -61,19 +78,34 @@ const destroyOne = async (request, response, next) => {
  */
 const destroyAll = async (request, response, next) => {
   try {
+    // Resource being created. We use this to check authorization later.
+    const THINGS = resources.THINGS;
+    // User instance requesting controller
+    const requestUser = request.user;
+
+    // Check if email matches req.user.email and set permission
+    const isAny = authorizations.can(requestUser.role).deleteAny(THINGS);
+
+    // Check if permission is grated
+    if (!isAny.granted) {
+      const error = new Error(`AUTHORIZATION ERROR: You are not authorized to ${request.method} resources on ${THINGS}.`);
+      error.statusCode = statusCodes.UNAUTHORIZED;
+      throw error;
+    }
+
     // Count the rows destroyed
     const countRowsDestroyed = await thingsService.destroyAll();
 
     // Check we have a row count to respond with
     if (!countRowsDestroyed) {
       const error = new Error(`CONTROLLER ERROR: Unable to destroy things.`);
-      error.statusCode = 500;
+      error.statusCode = statusCodes.INTERNAL_SERVER_ERROR;
       throw error;
     }
 
     // Destroy all response
-    const responseObject = response.status(200).json({
-      status: 200,
+    const responseObject = response.status(statusCodes.OK).json({
+      status: statusCodes.OK,
       message: `SUCCESS: Destroyed thing ${countRowsDestroyed} records.`,
       count: countRowsDestroyed,
     });
@@ -84,5 +116,5 @@ const destroyAll = async (request, response, next) => {
   }
 };
 
-export { destroyOne, destroyAll }
-export default { one: destroyOne, all: destroyAll }
+export { destroyOne, destroyAll };
+export default { one: destroyOne, all: destroyAll };
